requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
python-dotenv>=1.0.0
